Refs #628. More tests.
Former-commit-id: 2af66e468b02d36b1193a2c942df64df6846964d

+++ tests/unit/new/test_h5_rw.py
@@ -64,7 +64,7 @@ def test_create_data_node():
 
     node4 = DataSetNode(name)
     node4.add_attributes(attributes)
-    node4.set_value(np.ndarray([1, 2, 3.2]))
+    node4.set_value(np.ndarray([1, 2, 3.2], dtype=float))
 
     # Node 1 and 2 shall be same
     node2.match(node1)
